Fixing up and unify datatype to BigDecimal in DTOs

+++ src/main/kotlin/org/deblock/exercise/domain/dto/ToughJetSupplierDtoToFlight.kt
@@ -18,9 +18,9 @@ fun ToughJetSupplierDto.toFlight(supplier: String): Flight {
 
 fun ToughJetSupplierDto.calculateTotalPrice(): BigDecimal {
     val priceWithTax = basePrice + tax
-    val discountAmount = priceWithTax * (discount / 100)
+    val discountAmount = priceWithTax * BigDecimal(discount / 100)
     val totalPrice = priceWithTax - discountAmount
-    return BigDecimal(totalPrice).setScale(2, RoundingMode.HALF_UP)
+    return totalPrice.setScale(2, RoundingMode.HALF_UP)
 }
